Add a placeholder image for produtos without images

+++ src/utils/deleteImage.js
@@ -3,11 +3,13 @@ import fs from 'fs';
 
 const deleteImages = function (images) {
     images.forEach(img => {
-        fs.unlink(path.resolve('src/public/images/' + img), (err) => {
-            if (err) {
-                console.log(err);
-            }
-        });
+        if (img !== 'placeholder.png') {
+            fs.unlink(path.resolve('src/public/images/' + img), (err) => {
+                if (err) {
+                    console.log(err);
+                }
+            });
+        }
     });
 }
